backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
]
